Front End Update
Changed the shopping cart so that the subtotal reflects the real subtotal of the items in the cart. Adding an item to the cart now adds a spinner on the product quantity section, allowing the merchant to easily change the quantity of the item. If the quantity is set to 0, the merchant will be prompted if they wish to remove the item, in which case, the item is removed from the cart.