Front End Update #2
Most frames are now draggable in size, so you can resize the admin and merchant panels to something that seems more reasonable to you by dragging the arrow in the lower right hand corner. Most of the other updates are just fulfilling the rest of the use cases, such as the reward point management. All of this is still unhooked from the backend, hooking to the backend is my next job.