SMS Update
I've updated the frontend and backend to support sending SMS codes to the customer. For the Java portion, I have used Konnor's SMS generator for the bulk of the work. It's currently set to go through a random gmail account I set up, so it should work without you guys having to put in your own email.