Frontend Update
The customer's age is now stored with their shopping cart, so that it doesn't have to be re-entered everytime a minimum age product is purchased. A button to reset their age has also been implemented, which will clear their shopping cart of illegal items if their age is now below the minimum.